% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luq_anole_doc.R
\docType{data}
\name{luq_anole}
\alias{luq_anole}
\title{Anole Mark and Resight Study}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 543 rows and 16 columns.
}
\source{
{Reagan, D. 2015. Anole Mark and Resight Study ver 347700. Environmental Data Initiative. https://doi.org/10.6073/pasta/f9513a4dacda3880f2f6fe3774468bb1 (Accessed 2021-02-15).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-luq.5.347700}
}
\usage{
luq_anole
}
\description{
The population distribution research area of the LEF LTER is designed to gather
information on the distribution and abundance of key plant and animal species within
the forest under different disturbance regimes
}
\keyword{datasets}
