% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr_fish_doc.R
\docType{data}
\name{mcr_fish}
\alias{mcr_fish}
\title{MCR LTER: Coral Reef: Changes in the abundance of fish functional groups: Adam et al. 2014 Oecologia}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36 rows and 11 columns.
}
\source{
{ Moorea Coral Reef LTER, T. Adam, A. Brooks, and P. Edmunds. 2016. MCR LTER: Coral Reef: Changes in the abundance of fish functional groups: Adam et al. 2014 Oecologia ver 2. Environmental Data Initiative. https://doi.org/10.6073/pasta/2cea607bcfc63e41c5eee1da15e5a112 (Accessed 2021-02-15).}

\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-mcr.1041.2}
}
\usage{
mcr_fish
}
\description{
analyze the relationship between the cover of live and dead branching corals and changes in the
abundance of different functional groups of fishes following the loss of coral on the fore reef
of Moorea due to an outbreak of corallivorous crown-of-thorns sea stars (Acanthaster planci)
and a tropical cylcone.
}
\details{
A major ecosystem service provided by coral reefs is the provisioning of physical habitat for
other organisms, and consequently, many of the effects of climate change on coral reefs will
be mediated by their impacts on habitat structure
}
\keyword{datasets}
