% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_sediment_doc.R
\docType{data}
\name{pal_sediment}
\alias{pal_sediment}
\title{Sediment trap in nearshore waters collected during Palmer LTER (1992-95)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 335 rows and 12 columns.
}
\source{
{Palmer Station Antarctica LTER and D. Karl. 2017. Sediment trap in nearshore waters collected during Palmer LTER station season at Palmer Station Antarctica, 1992 - 1995. ver 3. Environmental Data Initiative. https://doi.org/10.6073/pasta/816f47788862f4dfc4437333e081bfb8 (Accessed 2021-02-15).}

\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-pal.46.3}
}
\usage{
pal_sediment
}
\description{
Particle fluxes to depth link the surface and mesopelagic realm and supply food to the benthos.
Sedimentation flux is typically measured with sediment traps of various designs.
Sedimentation at the PAL site of the West Antarctic Peninsula demonstrates extreme seasonality.
}
\keyword{datasets}
