% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloha.R
\name{say_aloha}
\alias{say_aloha}
\title{Says Aloha to a friend}
\usage{
say_aloha(name, print = TRUE)
}
\arguments{
\item{name}{a character}

\item{print}{a boolean, default = true}
}
\value{
(character) An aloha message
}
\description{
Says Aloha to a friend
}
\examples{
say_aloha("Lia")
}
